discord-bot/pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.4",
    "fastapi>=0.128.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "supabase>=2.27.2",
    "uvicorn>=0.40.0",
]
